==> 3_motifs.ipynb <==
import warnings
warnings.filterwarnings('ignore')

import pandas as pd 
import numpy as np
from statistics import mean
import re

proteins_with_functions = pd.read_csv("./Data/joined_tables.csv")

proteins_cleaned = proteins_with_functions.drop_duplicates(subset="structureId",keep="first")

## Sequence motif identification

I assumed that a good way to extract extra information from the data is to identify motifs that may be present in the sequences.

First, the first few amino acids at the beginning and end of each sequence are analysed. It is a simplified version of motif discovery which could have been executed on the whole sequence. This simple approach is reasonable, since in the polypeptide chains some signal peptides are often present at the N or C end of a protein. It made me think that to some extent it could be possible to distinguish protein types based on the first few amino acids in the sequence, as well as some short motifs at the end. 

The simplest way to check that was to slice first the first few amino acids, deduplicating sub-sequences, encoding them, and using them as a multi-class variable. To avoid increasing dimensionality, frequency encoding was initially attempted, however, this method produced many classes with a count = 1 and as a result it was not suitable for such case, so numerical encoding was used instead, which was eventually transformed into a one-hot encoding.

proteins_cleaned

proteins_dict = dict(zip(list(proteins_cleaned["classification"].value_counts().index), range(len(proteins_cleaned["classification"].value_counts().index.tolist()))))
proteins_cleaned["proteins_encoded"] = proteins_cleaned["classification"].map(proteins_dict)

proteins_cleaned["beg_seq"] = proteins_cleaned["sequence"].apply(lambda x: x[:10])
proteins_cleaned["end_seq"] = proteins_cleaned["sequence"].apply(lambda x: x[-10:])

value_dict = dict(zip(proteins_cleaned["beg_seq"].value_counts().keys().tolist(),
                      range(len(proteins_cleaned["beg_seq"].value_counts().index.tolist()))
                      ))


proteins_cleaned["start_seq_enc"] = proteins_cleaned["beg_seq"].map(value_dict)

value_dict = dict(zip(proteins_cleaned["end_seq"].value_counts().keys().tolist(),
                      range(len(proteins_cleaned["end_seq"].value_counts().index.tolist()))
                      ))


proteins_cleaned["end_seq_enc"] = proteins_cleaned["end_seq"].map(value_dict)

The RF was first used to make predictions only using the motifs as X variable and later with the entire combined dataset.

from sklearn.ensemble import RandomForestClassifier

df_size = len(proteins_cleaned)
train_size = int(len(proteins_cleaned)*0.7)

df = proteins_cleaned.sample(frac=1)

train = df[:train_size]
test = df[train_size:]

X_train = train[["start_seq_enc","end_seq_enc"]]
y_train = train["proteins_encoded"]

X_test = test[["start_seq_enc","end_seq_enc"]]
y_test = test["proteins_encoded"]

model = RandomForestClassifier()
model.fit(X_train,y_train)

y_pred = model.predict(X_test)

The results are quite promising, they indicate that there are some patterns in the data that enable to predict protein function based on the beginning and end motif.

from sklearn.metrics import accuracy_score
from sklearn.metrics import f1_score
from sklearn.metrics import precision_score
from sklearn.metrics import recall_score

print(f"accuracy: {accuracy_score(y_test, y_pred)}")
print(f"F1: {f1_score(y_test, y_pred, average='macro')}")
print(f"precision: {precision_score(y_test, y_pred, average='macro')}")
print(f"recall: {recall_score(y_test, y_pred, average='macro')}")

X_train = train.drop(columns=['structureId','chainId','sequence','classification','proteins_encoded', 'beg_seq', 'end_seq'])
y_train = train["proteins_encoded"]


X_test = test.drop(columns=['structureId','chainId','sequence','classification','proteins_encoded', 'beg_seq', 'end_seq'])
y_test = test["proteins_encoded"]

model = RandomForestClassifier()
model.fit(X_train,y_train)

The result haven't improved comparing to base model

y_pred = model.predict(X_test)

print(f"accuracy: {accuracy_score(y_test, y_pred)}")
print(f"F1: {f1_score(y_test, y_pred, average='macro')}")
print(f"precision: {precision_score(y_test, y_pred, average='macro')}")
print(f"recall: {recall_score(y_test, y_pred, average='macro')}")

motifs = df[["beg_seq","classification"]].groupby("classification").agg(list).reset_index()

motifs["len"]=motifs["beg_seq"].apply(len)

The sequences do not seem to be characteristic to the group, based on quick visual comparison. There are no typical patterns that characterises particular groups.

motifs

Another approach was to cluster similar sequences together. To simplify the process and make it more computationally efficient only 5 amino acids were sliced both from the beginning and the end of the sequence. 

proteins_cleaned["beg_seq"] = proteins_cleaned["beg_seq"].apply(lambda x: x[:5])
proteins_cleaned["end_seq"] = proteins_cleaned["end_seq"].apply(lambda x: x[:5])

proteins_cleaned

sequences_beg = pd.DataFrame({"beg_seq": proteins_cleaned['beg_seq'].value_counts().index, "count":proteins_cleaned["beg_seq"].value_counts().values})
sequences_end = pd.DataFrame({"end_seq": proteins_cleaned['end_seq'].value_counts().index, "count":proteins_cleaned["end_seq"].value_counts().values})

sequences_filtered_beg = sequences_beg[sequences_beg["count"]>=100]
sequences_filtered_end = sequences_end[sequences_end["count"]>=100]

Next step was to identify repetitive sequences within the dataset. From the sliced parts the occurrences of identical motifs were counted. It unfortunately resulted in relatively low values, with many unique sequences and numerous instances repeating only a few hundred times, which is not significant in a dataset of 100 000 instances. Despite this, the approach was tested to determine whether sequences could be grouped into similar motifs that might still be characteristic of specific proteins.  

Repetitive motifs were found in only around 1/3 of the instances in the dataset.

sequences_filtered_beg["count"].max()

sequences_filtered_beg["count"].sum()

sequences_filtered_end["count"].sum()

beg_motifs_df = pd.DataFrame(columns=np.array(sequences_filtered_beg["beg_seq"]), index=range(len(sequences_filtered_beg)))

Similarity between sequences was calculated using local sequence alignment from Biopython package. Then similarity matrix was constructed and calculated pairwise similarity. Finally a threshold (0,6) was established to define if sequences are similar to each other. 

On the resulting matrix connected component algorithm was executed to identify groups of similar motifs. They were labeled as one group. Motifs at the beginning and end of the sequences were identified and assigned corresponding labels. While this approach may not be ideal, as it could imply a numerical order, it was the simplest solution at this stage. Therefore, this encoding was used for the initial test.

# def jaccard_similarity(target,query):
#     set_a = set(target)
#     set_b = set(query)
#     score = len(set_a & set_b) / len(set_a| set_b)
#     return score

from Bio import Align
aligner = Align.PairwiseAligner()

for idx_col, column in enumerate(beg_motifs_df.columns):

    for idx_row, row in beg_motifs_df.iterrows():
        target = beg_motifs_df.columns[idx_col]
        query = beg_motifs_df.columns[idx_row]
        score = aligner.score(target, query) / (max(len(target),len(query)))

        beg_motifs_df.iloc[idx_row,idx_col] = score

beg_motifs_df

beg_motifs_df_binary = beg_motifs_df.copy()

for column in beg_motifs_df_binary.columns:
    beg_motifs_df_binary[column] = beg_motifs_df_binary[column].apply(lambda x: 1 if x>=0.6 else 0)

from scipy.sparse.csgraph import connected_components

n_components, labels = connected_components(csgraph=np.array(beg_motifs_df_binary), directed=False, return_labels=True)
n_components

beg_motifs_df_binary.index = beg_motifs_df_binary.columns
beg_motifs_df_binary["labels"] = labels
beg_motifs_df_binary["seq"] = beg_motifs_df_binary.index

beg_motifs_df_binary

beg_seq_dict = dict(zip(beg_motifs_df_binary["seq"], beg_motifs_df_binary["labels"]))

end_motifs_df = pd.DataFrame(columns=np.array(sequences_filtered_end["end_seq"]), index=range(len(sequences_filtered_end)))

for idx_col, column in enumerate(end_motifs_df.columns):

    for idx_row, row in end_motifs_df.iterrows():
        target = end_motifs_df.columns[idx_col]
        query = end_motifs_df.columns[idx_row]
        score = aligner.score(target, query) / (max(len(target),len(query)))

        end_motifs_df.iloc[idx_row,idx_col] = score

end_motifs_df_binary = end_motifs_df.copy()

for column in end_motifs_df_binary.columns:
    end_motifs_df_binary[column] = end_motifs_df_binary[column].apply(lambda x: 1 if x>=0.6 else 0)


n_components, labels = connected_components(csgraph=np.array(end_motifs_df_binary), directed=False, return_labels=True)


end_motifs_df_binary.index = end_motifs_df_binary.columns
end_motifs_df_binary["labels"] = labels
end_motifs_df_binary["seq"] = end_motifs_df_binary.index

end_seq_dict = dict(zip(end_motifs_df_binary["seq"], end_motifs_df_binary["labels"]))

n_components

proteins_cleaned["beg_motif"] = proteins_cleaned["beg_seq"].map(beg_seq_dict)
proteins_cleaned["end_motif"] = proteins_cleaned["end_seq"].map(end_seq_dict)

proteins_cleaned["beg_motif"].fillna(-1, inplace=True)
proteins_cleaned["end_motif"].fillna(-1, inplace=True)

Final test using Random Forest were performed and achieved very similar results to the base model. It is possible that the few first nucleotides do not provide enough information, cases belonging to any identified group constitute only around 25% of the dataset and as a result the information is probably insufficient. 

df = proteins_cleaned.sample(frac=1)

train = df[:train_size]
test = df[train_size:]

X_train = train.drop(columns=['structureId','chainId','sequence','classification','proteins_encoded', 'beg_seq', 'end_seq','start_seq_enc', 'end_seq_enc'])
y_train = train["proteins_encoded"]

X_test = test.drop(columns=['structureId','chainId','sequence','classification','proteins_encoded', 'beg_seq', 'end_seq','start_seq_enc', 'end_seq_enc'])
y_test = test["proteins_encoded"]

X_train.columns

model = RandomForestClassifier()
model.fit(X_train,y_train)

y_pred = model.predict(X_test)

print(f"accuracy: {accuracy_score(y_test, y_pred)}")
print(f"F1: {f1_score(y_test, y_pred, average='macro')}")
print(f"precision: {precision_score(y_test, y_pred, average='macro')}")
print(f"recall: {recall_score(y_test, y_pred, average='macro')}")

# import json

# with open('./Data/beg_motifs_dict.json', 'w') as fp:
#     json.dump(beg_seq_dict, fp)

# with open('./Data/end_motifs_dict.json', 'w') as fp:
#     json.dump(end_seq_dict, fp)
