==> 4_LSTM_model.ipynb <==
import warnings
warnings.filterwarnings('ignore')

import pandas as pd 
import numpy as np
import torch
import torch.nn as nn
import torch.optim as optim
import numpy as np

proteins_cleaned = pd.read_csv("./Data/joined_tables.csv")

Sequence characteristics were insufficient, and next step was to analyse the actual sequence information.

An LSTM model was initially selected, this model was the first choice because it seemed suitable, as this type of network can identify both shorter motifs and their relationships within the sequence.  

PyTorch was used to define the neural network, with various hyperparameters tested to achieve the best possible results.

# LSTM

First, the sequences were adjusted to have equal lengths by trimming longer sequences and applying padding to shorter ones. 

Different sequence lengths were tested, starting with 50 amino acids as the maximum due to computational limitations. Additionally, sequences of 20, 10, and 5 amino acids were testes. Surprisingly, the best results were achieved using a combination of the 5 initial and 5 amino acids from the end of a sequence 

def trimming_padding(sequence, leng, direction):

    if direction == "beg":
        new_seq = sequence[:leng]
        
        if len(new_seq) <= leng:
            fill = leng-len(new_seq)
            fill_seq = fill*"J"
            new_seq = new_seq + fill_seq
    else:
        new_seq = sequence[-leng:]
        
        if len(new_seq) <= leng:
            fill = leng-len(new_seq)
            fill_seq = fill*"J"
            new_seq = fill_seq + new_seq

    return new_seq

proteins_cleaned["trim_seq"] = proteins_cleaned["sequence"].apply(lambda x: trimming_padding(x,5, "beg"))
proteins_cleaned["trim_seq_end"] = proteins_cleaned["sequence"].apply(lambda x: trimming_padding(x,5,"end"))

Next, the encoding for the sequences had to be determined. To avoid adding too much dimensionality, encoding the letters of the alphabet in numerical order was considered. However, as it could introduce false relationships between elements within a sequence, one-hot encoding was ultimately chosen. While this method increased dimensionality, it still resulted in fewer dimensions than most embeddings.

aminoacids = ['A', 'C','D', 'E', 'F', 'G', 'H', 'I', 'K', 'L', 'M', 'N', 'O', 'P', 'Q', 'R','S', 'T', 'U', 'V', 'W', 'Y', 'X', 'B', 'Z']
encoding = [np.zeros(len(aminoacids)) for a in aminoacids]

for idx, enc in enumerate(encoding):
    enc[idx] = 1

aminoacids_dict = dict(zip(aminoacids,encoding))
aminoacids_dict.update({"J":np.zeros(len(aminoacids))})

aminoacids_dict["A"]

def map_seq(dict, sequence):

    array = np.zeros((len(sequence), 25))

    for idx, a in enumerate(sequence):
        array[idx] = dict[a]

    return array

Beginning and end sequences were combined together in one sequence

proteins_cleaned["seq_encoded"] = proteins_cleaned["trim_seq"].apply(lambda x: map_seq(aminoacids_dict,x))
proteins_cleaned["seq_encoded_end"] = proteins_cleaned["trim_seq_end"].apply(lambda x: map_seq(aminoacids_dict,x))

start_seq = np.array(proteins_cleaned["seq_encoded"].to_list())
end_seq = np.array(proteins_cleaned["seq_encoded_end"].to_list())

start_seq.shape

proteins_dict = dict(zip(list(proteins_cleaned["classification"].value_counts().index), range(len(proteins_cleaned["classification"].value_counts().index.tolist()))))

seq = np.concatenate((start_seq, end_seq), axis=1)

X = seq
y = np.array(proteins_cleaned["classification"].map(proteins_dict))

# np.save('./Data/x_lstm.npy', np.array(seq))
# np.save('./Data/y_lstm.npy', np.array(proteins_cleaned["classification"].map(proteins_dict)))

train_size = int(len(proteins_cleaned)*0.6)
test_size = int(len(proteins_cleaned)*0.2)

X_train = X[:train_size]
X_test = X[train_size:(train_size+test_size)]
X_val = X[(train_size+test_size):]

y_train = y[:train_size]
y_test = y[train_size:(train_size+test_size)]
y_val = y[(train_size+test_size):]

freq_dict = dict(zip(pd.DataFrame(y_train)[0].value_counts().index,
                     pd.DataFrame(y_train)[0].value_counts().values.tolist()))

dict_weights = dict()

weights_array = np.zeros(len(freq_dict))

for key, value in freq_dict.items():
    weight = 1-(value/len(y_train))
    weights_array[key] = weight

# np.save('./Data/weights.npy', weights_array)

X_train.shape

X_train_tensor_lstm = torch.tensor(X_train,dtype=torch.float32)
y_train_tensor_lstm = torch.tensor(np.array(y_train), dtype=torch.long)

X_test_tensor_lstm = torch.tensor(X_test,dtype=torch.float32)
y_test_tensor_lstm = torch.tensor(y_test, dtype=torch.long)

X_val_tensor_lstm = torch.tensor(X_val,dtype=torch.float32)
y_val_tensor_lstm = torch.tensor(y_val, dtype=torch.long)

Bidirectional as well as unidirectional networks were tested, with multiple values of numbers of layers and hidden size. 

The best results were achieved with bidirectional network, with 3 layers and hidden size 128.

class LSTMClassifier(nn.Module):
    def __init__(self, input_size, hidden_size, num_layers, output_size):
        super(LSTMClassifier, self).__init__()
        self.hidden_size = hidden_size
        self.num_layers = num_layers
        self.lstm = nn.LSTM(input_size, hidden_size, num_layers, batch_first=True, bidirectional=True)
        
        self.fc = nn.Linear(2*hidden_size, output_size)

    def forward(self, x):
        h0 = torch.zeros(2*self.num_layers, x.size(0), self.hidden_size)
        c0 = torch.zeros(2*self.num_layers, x.size(0), self.hidden_size)
        out, (hn, cn) = self.lstm(x, (h0, c0))
        out = self.fc(out[:, -1, :])
        return out

# UNDIRECTIONAL MODEL

# class LSTMClassifier(nn.Module):
#     def __init__(self, input_size, hidden_size, num_layers, output_size):
#         super(LSTMClassifier, self).__init__()
#         self.hidden_size = hidden_size
#         self.num_layers = num_layers
#         self.lstm = nn.LSTM(input_size, hidden_size, num_layers, batch_first=True)
        
#         self.fc = nn.Linear(hidden_size, output_size)

#     def forward(self, x):
#         h0 = torch.zeros(self.num_layers, x.size(0), self.hidden_size)
#         c0 = torch.zeros(self.num_layers, x.size(0), self.hidden_size)
#         out, (hn, cn) = self.lstm(x, (h0, c0))
#         out = self.fc(out[:, -1, :])
#         return out

torch.any(torch.isnan(X_train_tensor_lstm))

To prevent overfitting during the training loop, a simple version of early stopping was implemented. According to the definition: training was stopped when the loss on the training set began to increase while the loss on the validation set continued to decrease. Additionally, the best model was saved to be used for prediction in case it wasn't the one produced by the final epochs.

Both SGD and Adam optimizers were tested, with learning rates ranging from 0.001 to 0.1. The learning rate was increased to accelerate the learning process and decreased when fluctuations in the loss were observed during training.

input_size = 25
hidden_size = 128
num_layers = 3
output_size = 33

model_lstm = LSTMClassifier(input_size, hidden_size, num_layers, output_size)
criterion = nn.CrossEntropyLoss(weight=torch.tensor(weights_array,dtype=torch.float32))
optimizer = optim.Adam(model_lstm.parameters(), lr=0.01)

n_epochs = 15
batch_size = 128

test_loss_array = []
patience = 3
best_result = np.inf

for epoch in range(n_epochs):

    model_lstm.train()
    batches = len(X_train_tensor_lstm) // batch_size
    for batch in range(batches):
        
        optimizer.zero_grad()
        i = batch * batch_size

        X_batch = X_train_tensor_lstm[i:i+batch_size]
        y_batch = y_train_tensor_lstm[i:i+batch_size]

        y_pred = model_lstm(X_batch)

        loss = criterion(y_pred, y_batch)
        loss.backward()
        optimizer.step()

    model_lstm.eval()
    batches = len(X_test_tensor_lstm) // batch_size
    test_loss = 0
    correct = 0
    with torch.no_grad():  

        for batch in range(batches):
            i = batch * batch_size

            X_batch = X_test_tensor_lstm[i:i+batch_size]
            y_batch = y_test_tensor_lstm[i:i+batch_size]

            output = model_lstm(X_batch)
            test_loss += criterion(output, y_batch).item()

    test_loss /= len(y_test_tensor_lstm)
    test_loss_array.append(test_loss)

    if test_loss < best_result:
        torch.save(model_lstm.state_dict(), "./lstm.pth")

    print(f"Epoch: {epoch}, Train loss: {loss}, Test loss: {test_loss}")

    if len(test_loss_array)>patience+1:
        if not (any(x > test_loss_array[-1] for x in test_loss_array[len(test_loss_array)-patience-1:-1])):
            break

Best model was loaded from .pth file and used or a prediction on validation set

model_lstm.load_state_dict(torch.load("./lstm.pth"))

model_lstm.eval()
with torch.no_grad():
    predictions = model_lstm(X_val_tensor_lstm)
    predicted_labels = torch.argmax(predictions, dim=1)
    print("Predicted Labels:", predicted_labels)

results = pd.DataFrame({"pred":predicted_labels,"true":y_val})

from sklearn.metrics import accuracy_score
from sklearn.metrics import f1_score
from sklearn.metrics import precision_score
from sklearn.metrics import recall_score

The results still could be improved - i think the biggest change would be to use full sequences, but it is also possible that defining a function from the pure sequence is just way more complicated.


The results could still be improved. The most significant change would likely come from using full sequences. It is also possible that the sequence itself simply does not provide enough information and is not highly correlated with protein function.

y_test = results["true"]
y_pred = results["pred"]

print(f"accuracy: {accuracy_score(y_test, y_pred)}")
print(f"F1: {f1_score(y_test, y_pred, average='weighted')}")
print(f"precision: {precision_score(y_test, y_pred, average='weighted')}")
print(f"recall: {recall_score(y_test, y_pred, average='weighted')}")
