==> 7_additional_features.ipynb <==
import warnings
warnings.filterwarnings('ignore')

import pandas as pd 
import numpy as np
import scipy
import matplotlib
import seaborn as sns
from statistics import mean
import seaborn as sns
import matplotlib.pyplot as plt
import re

Finally, because the neural network results were not satisfying, the dataset with amino acid content was tested after being enriched with the proteins chemical properties and once again tested with Random Forest.

To achieve that protein properties from dataset (I) were combined with the preprocessed dataset (II created in 1_feature_ engineering)

proteins_with_functions = pd.read_csv("./Data/joined_tables.csv")

proteins_with_functions

proteins_data = pd.read_csv("./Data/pdb_data_no_dups.csv")

proteins_data

selected_columns= ["structureId","resolution","structureMolecularWeight","crystallizationTempK","densityMatthews","densityPercentSol","phValue"]

df = proteins_with_functions.merge(proteins_data[selected_columns].reset_index(), how="left", on="structureId")

df = df.drop_duplicates(subset="structureId",keep="first").reset_index()

df.drop(columns=["index","level_0"],inplace=True)

df

Information about motifs (from notebook 3_motifs) was added in this new data frame

import json

with open("./Data/beg_motifs_dict.json") as json_file:
    beg_motifs_dict = json.load(json_file)

with open("./Data/end_motifs_dict.json") as json_file:
    end_motifs_dict = json.load(json_file)

df["beg_motif"] = df["sequence"].apply(lambda x: x[:5]).map(beg_motifs_dict)
df["end_motif"] = df["sequence"].apply(lambda x: x[-5:]).map(end_motifs_dict)

df

Missing values were filled with mean and unidentified motifs with -1 

for column in df.columns[-8:-2]:
    df[column] = df[column].fillna(df.groupby('classification')[column].transform('mean'))

for column in df.columns[-2:]:
    df[column] = df[column].fillna(-1)
    df[column] = df[column].astype("int")

df

proteins_dict = dict(zip(list(df["classification"].value_counts().index), range(len(df["classification"].value_counts().index.tolist()))))
df["proteins_encoded"] = df["classification"].map(proteins_dict)

The results were checked with Random Forest. 
The improvements were small compared to the tests from 2_sequence_based_model notebook, but overall this enriched dataset enabled to achieve the best results. Even though many of the experiments didn’t beat the baseline model they enable to explore the data better and observe the patterns in it.

from sklearn.ensemble import RandomForestClassifier

df_size = len(df)
train_size = int(len(df)*0.7)

proteins_cleaned_cleaned = df.sample(frac=1)

train = proteins_cleaned_cleaned[:train_size]
test = proteins_cleaned_cleaned[train_size:]

X_train = train.drop(columns=['structureId','chainId','sequence','classification',"proteins_encoded"])
y_train = train["proteins_encoded"]

X_test = test.drop(columns=['structureId','chainId','sequence','classification',"proteins_encoded"])
y_test = test["proteins_encoded"]

weights_dict = dict(zip(
                         range(len(train["classification"].value_counts().index.tolist())), 
                         train["classification"].value_counts().values.tolist()))

dict_weights = dict()

count_val = len(X_train)
for key, value in weights_dict.items():
    weight = 1-(value/count_val)
    dict_weights.update({key:weight})

model = RandomForestClassifier(class_weight=dict_weights)
model.fit(X_train,y_train)

y_pred = model.predict(X_test)

Results improved compairing to the results of the sequence itself

from sklearn.metrics import accuracy_score
from sklearn.metrics import f1_score
from sklearn.metrics import precision_score
from sklearn.metrics import recall_score


print(f"accuracy: {accuracy_score(y_test, y_pred)}")
print(f"F1: {f1_score(y_test, y_pred, average='macro')}")
print(f"precision: {precision_score(y_test, y_pred, average='macro')}")
print(f"recall: {recall_score(y_test, y_pred, average='macro')}")

# df.to_csv("./Data/additional_features_dataset.csv")

df
