==> 8_combined_dataset_xgb.ipynb <==
import warnings
warnings.filterwarnings('ignore')

import pandas as pd

from sklearn.metrics import accuracy_score
from sklearn.metrics import f1_score
from sklearn.metrics import precision_score
from sklearn.metrics import recall_score

Because the best results so far were achieved using a combined dataset with feature engineering and a Random Forest model, it allows to think that tree-based models may be the most optimal for this problem.

Because if that XGBoost was selected for further testing.

### Preprocessing

proteins_with_add = pd.read_csv("./Data/additional_features_dataset.csv")

proteins_with_add

proteins_with_add = proteins_with_add.drop_duplicates(subset="structureId",keep="first").reset_index()

proteins_with_add.drop(columns=["Unnamed: 0" , "index"],inplace=True)

proteins_with_add

### Model

from xgboost import XGBClassifier

df_size = len(proteins_with_add)
train_size = int(len(proteins_with_add)*0.7)

proteins_cleaned_cleaned = proteins_with_add.sample(frac=1)

train = proteins_cleaned_cleaned[:train_size]
test = proteins_cleaned_cleaned[train_size:]

X_train = train.drop(columns=['structureId','chainId','sequence','classification',"proteins_encoded"])
y_train = train["proteins_encoded"]

X_test = test.drop(columns=['structureId','chainId','sequence','classification',"proteins_encoded"])
y_test = test["proteins_encoded"]

from sklearn.utils.class_weight import compute_sample_weight

sample_weights = compute_sample_weight(class_weight='balanced', y=y_train)

def test_result(model, X_test, y_test):
    y_pred = model.predict(X_test)
    print(f"accuracy: {accuracy_score(y_test, y_pred)}")
    print(f"F1: {f1_score(y_test, y_pred, average='macro')}")
    print(f"precision: {precision_score(y_test, y_pred, average='macro')}")
    print(f"recall: {recall_score(y_test, y_pred, average='macro')}")

Grid search was performed to find the optimal hyperparameter values: n_estimators parameter was fixed at 100, as lower values were unlikely to obtain satisfactory results, while higher values significantly increased execution time.

from sklearn.model_selection import GridSearchCV

params = {
    'n_estimators': [100],
    'max_depth': [5, 10],
    'learning_rate': [0.1, 0.01],
    'subsample': [0.5, 1]
}

model = XGBClassifier(booster='gbtree',
                      tree_method='hist',
                      random_state=42)

grid_search = GridSearchCV(model, params, cv=5, scoring='accuracy')

grid_search.fit(X_train, y_train, sample_weight = sample_weights)

print(f"Best params: {grid_search.best_params_}, acc: {grid_search.best_score_}")

Based on the best model 3 more tests were performed:
1. With same number of estimators but higher depth (15)
2. With increased number estimators (1000) but same depth (10)
3. With increased number of estimators and increased depth 

model = XGBClassifier(
    booster='gbtree',
    n_estimators=100,
    max_depth=15,
    learning_rate=0.1,
    subsample=1,
    num_parallel_tree=1,
    tree_method='hist',
    random_state=42
)

model.fit(X_train, y_train, sample_weight = sample_weights)
print("")

test_result(model, X_test, y_test)

model = XGBClassifier(
    booster='gbtree',
    n_estimators=1000,
    max_depth=10,
    learning_rate=0.1,
    subsample=1,
    num_parallel_tree=1,
    tree_method='hist',
    random_state=42
)

model.fit(X_train, y_train, sample_weight = sample_weights)
print("")

test_result(model, X_test, y_test)

model = XGBClassifier(
    booster='gbtree',
    n_estimators=1000,
    max_depth=15,
    learning_rate=0.1,
    subsample=1,
    num_parallel_tree=1,
    tree_method='hist',
    random_state=42
)

model.fit(X_train, y_train, sample_weight = sample_weights)
print("")

test_result(model, X_test, y_test)

The final test showed no improvement, so the optimal hyperparameters remained:

- n_estimators=1000,
- max_depth=10,
- learning_rate=0.1,
- subsample=1,
- num_parallel_tree=1

With a result of: 
- accuracy: 0.85
- F1: 0.83
- precision: 0.86
- recall: 0.80
