==> 9_results.ipynb <==
The best results were achieved using a Random Forest model on the combined dataset of protein properties and sequence statistics. The amino acid content in the polypeptide was particularly useful for accurate prediction. This approach improved the prediction performance significantly.

The results achieved with Random forest are:
- On the protein properties dataset (I): accuracy: 0.63, F1: 0.60 
- On the protein sequences dataset (II), after feature engineering: accuracy: 0.76, F1: 0.76  
- On the combined dataset: accuracy: 0.82, F1: 0.79  

PyTorch networks were performing worse than RF and the best results were achieved when the protein properties were enriched by the LSTM prediction. As input to the LSTM, the beginning and end of the sequence were used, and later the predictions from LSTM were then combined with the protein properties dataset to enhance it with additional protein characteristics. This 2-step prediction enable to achieve:
- accuracy: 0.62, F1: 0.61

While the results on the Neural Networks separately were:
- LSTM accuracy: 0.55, F1: 0.55
- NN Classifier accuracy: 0.50, F1: 0.48


While the Random Forest model achieved decent results, there was still room for improvement.
Because of that Gradient Boosting was also tested on the dataset. After optimizing the hyperparameters it slightly improved the final prediction accuracy with XGBoostClassifier achieving following results:

- accuracy: **0.85**
- F1: **0.83**


On the other hand it cold be also beneficial to further enrich the dataset with more polypeptide data - the protein properties dataset (I) had many NaN values, which were filled with the mean of the group, which in some cases could be inaccurate and disrupt the natural correlations in the data.
For example longer sequences tend to have higher molecular weight, but this correlation could be lost in the situation where a long sequence would be filled with a low mean molecular mass value calculated based on shorter sequences. 
So it would be even better to collect more data instead of only handling the missing values.
